refactored trainer into local and distributed classes, cleaned logic and creater roundTracker class
