+++ mobile-browser-based-version/tests/api.test.ts
@@ -23,14 +23,7 @@ describe('api test', () => { // the tests container
   it('getRound', async () => {
     await api.connect(task, user)
     const resp = await api.getRound(task, user)
-    expect(resp.data.round).equal(0)
-    await api.disconnect(task, user)
-  })
-
-  it('getRound at server init', async () => {
-    await api.connect(task, user)
-    const resp = await api.getRound(task, user)
-    expect(resp.data.round).equal(0)
+    expect(resp.data.round).greaterThanOrEqual(0) // Since the server you are running might have trained and round > 0
     await api.disconnect(task, user)
   })
 })
